openai1/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
Django==4.1.7
filelock==3.10.0
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.2
platformdirs==3.1.1
pygame==2.2.0
python-dotenv==1.0.0
requests==2.28.2
sqlparse==0.4.3
tqdm==4.65.0
urllib3==1.26.15
virtualenv==20.21.0
Werkzeug==2.2.3
yarl==1.8.2
